+++ common/src/main/java/com/yanny/ali/compatibility/jei/JeiGameplayLoot.java
@@ -1,14 +1,19 @@
 package com.yanny.ali.compatibility.jei;
 
+import com.yanny.ali.api.Rect;
 import com.yanny.ali.compatibility.common.GameplayLootType;
+import com.yanny.ali.compatibility.common.GenericUtils;
 import com.yanny.ali.registries.LootCategory;
+import mezz.jei.api.gui.builder.ITooltipBuilder;
 import mezz.jei.api.gui.drawable.IDrawable;
 import mezz.jei.api.gui.ingredient.IRecipeSlotDrawable;
 import mezz.jei.api.gui.widgets.IRecipeExtrasBuilder;
 import mezz.jei.api.gui.widgets.IRecipeWidget;
 import mezz.jei.api.helpers.IGuiHelper;
 import mezz.jei.api.recipe.RecipeType;
+import net.minecraft.client.gui.navigation.ScreenPosition;
 import net.minecraft.network.chat.Component;
+import org.jetbrains.annotations.NotNull;
 import oshi.util.tuples.Pair;
 
 import java.util.List;
@@ -20,12 +25,33 @@ public class JeiGameplayLoot extends JeiBaseLoot<GameplayLootType, String> {
 
     @Override
     Pair<List<IRecipeWidget>, List<IRecipeSlotDrawable>> getWidgets(IRecipeExtrasBuilder builder, GameplayLootType recipe) {
-        Component text = Component.translatableWithFallback("ali/loot_table/" + recipe.id().substring(1), recipe.id());
-        return new Pair<>(List.of(createTextWidget(text, 0, 0, false)), List.of());
+        String key = "ali/loot_table/" + recipe.id().substring(1);
+        Component text = GenericUtils.ellipsis(key, recipe.id(), CATEGORY_WIDTH);
+        Component fullText = Component.translatableWithFallback(key, recipe.id());
+        Rect rect = new Rect(0, 0, CATEGORY_WIDTH, 8);
+        return new Pair<>(List.of(
+                createTextWidget(text, 0, 0, false),
+                new TooltipWidget(fullText, rect)
+        ), List.of());
     }
 
     @Override
     int getYOffset(GameplayLootType recipe) {
         return 10;
+    }
+
+    private record TooltipWidget(Component component, Rect rect) implements IRecipeWidget {
+        @NotNull
+        @Override
+        public ScreenPosition getPosition() {
+            return new ScreenPosition(rect().x(), rect.y());
+        }
+
+        @Override
+        public void getTooltip(ITooltipBuilder tooltip, double mouseX, double mouseY) {
+            if (rect.contains((int) mouseX, (int) mouseY)) {
+                tooltip.add(component);
+            }
+        }
     }
 }